+++ HISTORY.rst
@@ -5,6 +5,7 @@ UNRELEASED
 ++++++++++
 
 - Add initial support for OAuth Mutual TLS (draft-ietf-oauth-mtls)
+- Removed outdated LinkedIn Compliance Fixes
 
 v1.3.0 (6 November 2019)
 ++++++++++++++++++++++++